Rename package to tools.kotlin.dockerutils.imageio
